fix: signin, signup, changePassword type
