--- apps/admin/src/shared/model/changePWForm.ts
@@ -1,8 +1,0 @@
-export interface ChangePasswordProps {
-  email: string;
-  password: string;
-}
-
-export interface ChangePasswordForm extends ChangePasswordProps {
-  passwordCheck: string;
-}
