+++ apps/admin/src/shared/model/changePassword.ts
@@ -1,0 +1,8 @@
+import type { StepAuthCodeForm } from "./signup";
+
+export interface ChangePasswordProps {
+  email: string;
+  password: string;
+}
+
+export type ChangePassword_StepAuthCodeForm = Omit<StepAuthCodeForm, "name">;
