+++ apps/admin/src/shared/model/signup.ts
@@ -1,0 +1,15 @@
+export interface SignupFormProps {
+  email: string;
+  password: string;
+  name: string;
+}
+export interface StepAuthCodeForm {
+  email: string;
+  authcode: string;
+  name: string;
+}
+
+export interface StepPasswordForm {
+  password: string;
+  passwordCheck: string;
+}
